feat(M): recreate initial migration script
